:package: build: upgrade flutter_mongo_stitch_platform_interface :package: build: change sdk minimum version to support null safety :recycle: refactor: update in dart files to support null safety

+++ lib/flutter_mongo_stitch_web.dart
@@ -2,7 +2,6 @@ library flutter_mongo_stitch_web;
 
 import 'dart:convert';
 
-import 'package:flutter/material.dart';
 import 'package:flutter_mongo_stitch_platform_interface/flutter_mongo_stitch_platform_interface.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
@@ -14,7 +13,7 @@ import 'src/interop.dart';
 ///
 /// This class implements the `package:flutter_mongo_stitch` functionality for the web.
 class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
-  MyMongoClient _mongoClient;
+  late MyMongoClient _mongoClient;
   bool _injected = false;
 
   static void registerWith(Registrar registrar) async {
@@ -45,9 +44,9 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future insertDocument({
-    @required String collectionName,
-    @required String databaseName,
-    @required Map<String, Object> data,
+    required String collectionName,
+    required String databaseName,
+    required Map<String, Object?> data,
   }) async {
     var id =
         await _mongoClient.insertDocument(databaseName, collectionName, data);
@@ -56,9 +55,9 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future insertDocuments({
-    @required String collectionName,
-    @required String databaseName,
-    @required List<String> list,
+    required String collectionName,
+    required String databaseName,
+    required List<String> list,
   }) async {
     var map =
         await _mongoClient.insertDocuments(databaseName, collectionName, list);
@@ -68,12 +67,19 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future findDocuments(
-      {String collectionName,
-      String databaseName,
+      {String? collectionName,
+      String? databaseName,
       dynamic filter,
-      String projection,
-      int limit,
-      String sort}) async {
+      String? projection,
+      int? limit,
+      String? sort}) async {
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
     var list =
         await _mongoClient.findDocuments(databaseName, collectionName, filter);
     return Future.value(list);
@@ -81,11 +87,18 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future findFirstDocument(
-      {String collectionName,
-      String databaseName,
+      {String? collectionName,
+      String? databaseName,
       dynamic filter,
-      String projection}) async {
+      String? projection}) async {
     //todo, ADD: final String projection = call.arguments['projection'];
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
 
     var list =
         await _mongoClient.findDocument(databaseName, collectionName, filter);
@@ -94,7 +107,14 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future deleteDocument(
-      {String collectionName, String databaseName, dynamic filter}) async {
+      {String? collectionName, String? databaseName, dynamic filter}) async {
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
     String resultString =
         await _mongoClient.deleteDocument(databaseName, collectionName, filter);
     Map<String, dynamic> map = json.decode(resultString);
@@ -104,7 +124,14 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future deleteDocuments(
-      {String collectionName, String databaseName, dynamic filter}) async {
+      {String? collectionName, String? databaseName, dynamic filter}) async {
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
     String resultString = await _mongoClient.deleteDocuments(
         databaseName, collectionName, filter);
     Map<String, dynamic> map = json.decode(resultString);
@@ -114,7 +141,15 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future countDocuments(
-      {String collectionName, String databaseName, dynamic filter}) async {
+      {String? collectionName, String? databaseName, dynamic filter}) async {
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
+
     var size =
         await _mongoClient.countDocuments(databaseName, collectionName, filter);
     return Future.value(size);
@@ -122,10 +157,26 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future updateDocument(
-      {String collectionName,
-      String databaseName,
-      String filter,
-      String update}) async {
+      {String? collectionName,
+      String? databaseName,
+      String? filter,
+      String? update}) async {
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
+
+    if (filter == null) {
+      throw Exception("filter should not be null.");
+    }
+
+    if (update == null) {
+      throw Exception("update should not be null.");
+    }
+
     String resultString = await _mongoClient.updateDocument(
         databaseName, collectionName, filter, update);
     Map<String, dynamic> map = json.decode(resultString);
@@ -135,10 +186,26 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
 
   @override
   Future updateDocuments(
-      {String collectionName,
-      String databaseName,
-      String filter,
-      String update}) async {
+      {String? collectionName,
+      String? databaseName,
+      String? filter,
+      String? update}) async {
+    if (databaseName == null) {
+      throw Exception("databaseName should not be null.");
+    }
+
+    if (collectionName == null) {
+      throw Exception("collectionName should not be null.");
+    }
+
+    if (filter == null) {
+      throw Exception("filter should not be null.");
+    }
+
+    if (update == null) {
+      throw Exception("update should not be null.");
+    }
+
     String resultString = await _mongoClient.updateDocuments(
         databaseName, collectionName, filter, update);
     Map<String, dynamic> map = json.decode(resultString);
@@ -217,14 +284,17 @@ class FlutterMongoStitchPlugin extends FlutterMongoStitchPlatform {
   /// =====
 
   @override
-  Future callFunction(String name, {List args, int requestTimeout}) async {
+  Future callFunction(String name, {List? args, int? requestTimeout}) async {
+    if (args == null) {
+      throw Exception("args  should not be null.");
+    }
     var result = _mongoClient.callFunction(name, args); //, timeout);
     return Future.value(result);
   }
 
   @override
   Future setupWatchCollection(String collectionName, String databaseName,
-      {List<String> ids, bool asObjectIds, String filter}) {
+      {List<String>? ids, bool? asObjectIds, String? filter}) {
     if (filter == null) {
       if (ids == null || ids.isEmpty) {
         _mongoClient.setupWatchCollection(databaseName, collectionName);
